文件管理及记忆化
# message auto-generated for no-merge-commit merge:
merge ******_project_manager into master

文件管理及记忆化，第二部分代码合入

Created-by: insightDeveloper

Match-id-5bb02714d4e48fb41f0310bc0d9851d4626d4b2a

+++ framework/src/components/MenuTree/MenuTree.vue
@@ -2,31 +2,75 @@
 import type { TreeNodeType } from './types';
 import LocalIcon from '@/components/icons/loaclImport_icon.vue';
 import DeletePopConfirm from '@/components/MenuTree/DeletePopConfirm.vue';
+import EditableText from '@/components/MenuTree/EditableText.vue';
+import AddIcon from '@/components/icons/cross_icon.vue';
+import {ref, watch} from 'vue';
+import {useDataSources} from '@/stores/dataSource';
+import { LOCAL_HOST, PORT } from '@/centralServer/websocket/defs';
 import useWatchTranslation from '@/hooks/useWatchTranslation';
+import ResourceDialog from '@/components/ResourceDialog.vue';
 
-const [DeleteAll, DeleteItem] = useWatchTranslation(['Delete All', 'Delete Item']);
+const showModal = ref(false);
+const projectName = ref('');
+const activateNodeId = ref(0);
+
+
+const [DeleteAll, DeleteItem, Cancel, Confirm] = useWatchTranslation(['Delete All', 'Delete Item', 'Cancel', 'Confirm']);
 const props = defineProps<{
     dataSource: TreeNodeType[];
     isDarkTheme: boolean;
 }>();
 
+watch(
+    () => [useDataSources().lastDataSource, props.dataSource],
+    ()=>{
+      const idx = props.dataSource.findIndex((item) => item.label === useDataSources().lastDataSource?.projectName);
+      if (idx === -1) {
+        activateNodeId.value = -1;
+        return;
+      }
+      const project = props.dataSource[idx];
+      activateNodeId.value = project.id;
+    }
+);
+
+const handleNodeClick = (data:any, node: any) => {
+  if (node.level === 1) {
+    const dataSource = { remote: LOCAL_HOST, port: PORT, projectName: data.label, dataPath: [] };
+    useDataSources().confirm(dataSource, false);
+  }
+};
+
+function addRemoteUnderProject(node:any, e: MouseEvent) {
+  e.stopPropagation();
+  projectName.value = node.data.label;
+  showModal.value = true;
+}
+
 </script>
 <template>
     <div class="menu-tree">
-        <el-tree :data="props.dataSource" node-key="id" :default-expand-all="true">
+        <el-tree :data="props.dataSource" node-key="id" :default-expand-all="true" :expand-on-click-node="false" :current-node-key="2" @node-click="handleNodeClick">
             <template #default="{ node, data }">
-                <div class="contentNode">
+                <div :class="['contentNode', {activateNode:data.id === activateNodeId}]">
                     <span class="contentBody">
-                        <LocalIcon v-if="node.level === 1" />
-                        <span v-if="node.level === 1" class="contentText">
-                            {{ node.label }}
-                        </span>
-                        <el-tooltip v-if="node.level === 2" :content="node.label" :effect="isDarkTheme ? 'light' : 'dark'">
+                        <LocalIcon v-if="node.level === 1" style="flex: none"/>
+                        <el-tooltip :content="node.label" v-if="node.level === 1" :effect="isDarkTheme ? 'light' : 'dark'">
+                           <EditableText :tree-node="node" :key="data.id + data.label"></EditableText>
+                        </el-tooltip>
+                        <el-tooltip v-if="node.level === 2" :node-key="node.label" :content="node.label" :effect="isDarkTheme ? 'light' : 'dark'">
                             <span class="contentNodeText">
                                 {{ node.label }}
                             </span>
                         </el-tooltip>
                     </span>
+                    <el-tooltip v-if="node.level === 1" content="Import Data" :effect="isDarkTheme ? 'light' : 'dark'">
+                      <span class="addIcon">
+                        <el-icon class="icon-button" @click.stop="addRemoteUnderProject(node, $event)">
+                          <AddIcon />
+                        </el-icon>
+                      </span>
+                    </el-tooltip>
                     <el-tooltip v-if="node.level === 1" :content="DeleteAll" :effect="isDarkTheme ? 'light' : 'dark'">
                             <span class="deleteIcon">
                               <DeletePopConfirm :data="data" :node="node" :is-delete-all="node.level === 1"/>
@@ -41,6 +85,7 @@ const props = defineProps<{
                 </div>
             </template>
         </el-tree>
+        <ResourceDialog v-model:showModal=showModal :project-name="projectName"></ResourceDialog>
     </div>
 </template>
 <style scoped>
@@ -52,10 +97,13 @@ const props = defineProps<{
     --el-tree-node-hover-bg-color: var(--color-border-hover);
     color: var(--dataPath-color) !important;
 }
+::v-deep(.activateNode) {
+  background-color:var(--treeNode-onclick);
+}
 
 ::v-deep(.contentNode) {
     position: relative;
-    display: contents;
+    display: flex;
     user-select: none;
     justify-content: space-between;
     flex: 1;
@@ -68,7 +116,9 @@ const props = defineProps<{
     align-items: center;
     width: 0;
     flex: 1 1 auto;
-    margin-right: 20px;
+    margin-right: 50px;
+    overflow: hidden;
+    text-overflow: ellipsis;
 }
 
 ::v-deep(.treeDeleteIcon) {
@@ -92,6 +142,14 @@ const props = defineProps<{
     font-weight: 400;
     overflow: hidden;
     white-space: nowrap;
+}
+
+::v-deep(.addIcon) {
+  flex: 0 0 auto;
+  position: absolute;
+  right: 20px;
+  height: 16px;
+  margin-top: 2px;
 }
 
 ::v-deep(.deleteIcon) {
